Fixed warnings. JIRA: CXX-4704
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@63782 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/data_loaders/genbank/impl/dispatcher.hpp
@@ -116,7 +116,7 @@ public:
     void LoadStates(CReaderRequestResult& result,
                     const TIds ids, TLoaded& loaded, TStates& ret);
  
-    static bool SetBlobState(int i,
+    static bool SetBlobState(size_t i,
                              CReaderRequestResult& result,
                              const TIds& ids, TLoaded& loaded,
                              TStates& ret);
